Move extraction symlink tests to test_reader.cc
